Update Christine's Analysis.md
INSIGHT -

OUR FIRST QUESTION is the most important. HAS the rate of part-time employment changed over time.

We need to find data from different decades to answer our second question: is this "insidious" problem really new?

If the problem isn't new then we have teeth for our argument to apply FTE standards and require retailers to offer health benefits to part-time personnel.

IF THERE HAS BEEN A SIGNIFICANT CHANGE and this "insidious" problem is NEW, then we have to ask why? What is changing in the retail workforce OR what is changing in general in  the labor market.